Miscellaneous Tweaks
Move speed, menu behaviour, etc.

+++ Player.hpp
@@ -106,9 +106,9 @@ private:
     const int PLAYER_SIZE = 16;
     const int MOVE_SPEED = 50;
 
-    const int MOVESCALE_BASE = 0.7;
-    const int MOVESCALE_1 = 0.9;
-    const int MOVESCALE_2 = 1.1;
+    const float MOVESCALE_BASE = 0.7;
+    const float MOVESCALE_1 = 0.9;
+    const float MOVESCALE_2 = 1.1;
 
     const int ROTATE_SPEED = 90;
     const int TURRET_ROTATE_SPEED = 110;
